+++ platform-sdk/swirlds-common/src/main/java/com/swirlds/common/wiring/schedulers/internal/NoOpTaskScheduler.java
@@ -16,7 +16,7 @@
 
 package com.swirlds.common.wiring.schedulers.internal;
 
-import com.swirlds.common.wiring.model.internal.TraceableWiringModel;
+import com.swirlds.common.wiring.model.internal.standard.TraceableWiringModel;
 import com.swirlds.common.wiring.schedulers.TaskScheduler;
 import com.swirlds.common.wiring.schedulers.builders.TaskSchedulerType;
 import com.swirlds.common.wiring.wires.input.BindableInputWire;
