+++ platform-sdk/swirlds-common/src/main/java/com/swirlds/common/wiring/wires/input/NoOpInputWire.java
@@ -16,7 +16,7 @@
 
 package com.swirlds.common.wiring.wires.input;
 
-import com.swirlds.common.wiring.model.internal.TraceableWiringModel;
+import com.swirlds.common.wiring.model.internal.standard.TraceableWiringModel;
 import com.swirlds.common.wiring.schedulers.TaskScheduler;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.function.Consumer;
