+++ platform-sdk/swirlds-common/src/main/java/com/swirlds/common/wiring/wires/output/NoOpOutputWire.java
@@ -16,7 +16,7 @@
 
 package com.swirlds.common.wiring.wires.output;
 
-import com.swirlds.common.wiring.model.internal.TraceableWiringModel;
+import com.swirlds.common.wiring.model.internal.standard.TraceableWiringModel;
 import com.swirlds.common.wiring.transformers.AdvancedTransformation;
 import com.swirlds.common.wiring.wires.SolderType;
 import com.swirlds.common.wiring.wires.input.InputWire;
