+++ platform-sdk/swirlds-common/src/main/java/com/swirlds/common/wiring/wires/output/internal/ForwardingOutputWire.java
@@ -16,7 +16,7 @@
 
 package com.swirlds.common.wiring.wires.output.internal;
 
-import com.swirlds.common.wiring.model.internal.TraceableWiringModel;
+import com.swirlds.common.wiring.model.internal.standard.TraceableWiringModel;
 import com.swirlds.common.wiring.wires.output.OutputWire;
 import edu.umd.cs.findbugs.annotations.NonNull;
